toolkit/mozapps/downloads/tests/chrome/test_close_download_manager.xul

<?xml version="1.0"?>
<!--
/* ***** BEGIN LICENSE BLOCK *****
 * Version: MPL 1.1/GPL 2.0/LGPL 2.1
 *
 * The contents of this file are subject to the Mozilla Public License Version
 * 1.1 (the "License"); you may not use this file except in compliance with
 * the License. You may obtain a copy of the License at
 * http://www.mozilla.org/MPL/
 *
 * Software distributed under the License is distributed on an "AS IS" basis,
 * WITHOUT WARRANTY OF ANY KIND, either express or implied. See the License
 * for the specific language governing rights and limitations under the
 * License.
 *
 * The Original Code is Download Manager UI Test Code.
 *
 * The Initial Developer of the Original Code is
 * Anoop Saldanha <poonaatsoc@gmail.com>
 *
 * Portions created by the Initial Developer are Copyright (C) 2008
 * the Initial Developer. All Rights Reserved.
 *
 * Contributor(s):
 *
 * Alternatively, the contents of this file may be used under the terms of
 * either the GNU General Public License Version 2 or later (the "GPL"), or
 * the GNU Lesser General Public License Version 2.1 or later (the "LGPL"),
 * in which case the provisions of the GPL or the LGPL are applicable instead
 * of those above. If you wish to allow use of your version of this file only
 * under the terms of either the GPL or the LGPL, and not to allow others to
 * use your version of this file under the terms of the MPL, indicate your
 * decision by deleting the provisions above and replace them with the notice
 * and other provisions required by the GPL or the LGPL. If you do not delete
 * the provisions above, a recipient may use your version of this file under
 * the terms of any one of the MPL, the GPL or the LGPL.
 *
 * ***** END LICENSE BLOCK ***** */

/**
 * This test basically checks if the download manager
 * closes when you press the esc key and accel + w.
 */
-->

<window title="Download Manager Test"
        onload="runTest();"
        xmlns:html="http://www.w3.org/1999/xhtml"
        xmlns="http://www.mozilla.org/keymaster/gatekeeper/there.is.only.xul">

  <script type="application/javascript"
          src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"/>
  <script type="application/javascript"
          src="chrome://mochikit/content/tests/SimpleTest/EventUtils.js"/>
  <script type="application/javascript"
          src="utils.js"/>

  <script type="application/javascript">

  <![CDATA[

const dmui = getDMUI();

function testCloseDMWithEscKey(aWin)
{
  function dmWindowClosedListener() {
    aWin.removeEventListener("unload", dmWindowClosedListener, false);
    ok(!dmui.visible, "DMUI closes with ESC key");
    dmui.show();
  }
  aWin.addEventListener("unload", dmWindowClosedListener, false);

  synthesizeKey("VK_ESCAPE", { }, aWin);
}

function testCloseDMWithAccelKey(aWin)
{
  function dmWindowClosedListener() {
    aWin.removeEventListener("unload", dmWindowClosedListener, false);
    ok(!dmui.visible, "DMUI closes with accel + w");
    setCleanState();
    SimpleTest.finish();
  }
  aWin.addEventListener("unload", dmWindowClosedListener, false);

  synthesizeKey("w", { accelKey: true }, aWin);
}

function runTest()
{
  if (!dmui) {
    todo(false, "skip test for toolkit download manager UI");
    return;
  }

  const DLMGR_UI_DONE = "download-manager-ui-done";

  // Close the DM UI if it is already open
  let dm_win = getDMWindow();
  if (dm_win) dm_win.close();

  let testPhase = 0;
  // Specify an observer that will be notified when the dm has been rendered on screen
  let obs = Cc["@mozilla.org/observer-service;1"].
            getService(Ci.nsIObserverService);
  let testObs = {
    observe: function(aSubject, aTopic, aData) {
      SimpleTest.waitForFocus(function () { closeDM(aSubject) }, aSubject);
    }
  };

  function closeDM(win) {
    switch(testPhase++) {
      case 0:
        testCloseDMWithEscKey(win);
        break;
      case 1:
        obs.removeObserver(testObs, DLMGR_UI_DONE);
        testCloseDMWithAccelKey(win);
    }
  }
  obs.addObserver(testObs, DLMGR_UI_DONE, false);

  dmui.show();

  SimpleTest.waitForExplicitFinish();
}

  ]]>
  </script>

  <body xmlns="http://www.w3.org/1999/xhtml">
    <p id="display"></p>
    <div id="content" style="display: none"></div>
    <pre id="test"></pre>
  </body>
</window>
